fix race condition around unlink during shutdown
the server now always links when it comes back up. this means any data that
comes in while no one is attached will be dropped, but this guarantee only
applied up to 64k of data anyway. now unlink will always wait for there to be
a linked process to be kill.

Signed-off-by: Alex Suraci <asuraci@pivotallabs.com>