don't stop processes through wsh
when a container runs out of memory, it's possible that wshd itself is paused,
and so stop.sh would just hang.

instead, we stop all processes from the outside by going over the container's
cgroup's list of tasks. every second we send TERM to all processes (except for
wshd), and after 10 seconds that increases to KILL.

we're careful to send the signals to all proceses in the container, not just
the parent's immediate children. added integration tests to ensure this case.

Signed-off-by: Chris Adams <cadams@pivotallabs.com>