actually implement stdin streaming; replace iomux
iomux-spawn/link has been replaced with the much simpler and much easier to
test iodaemon. rather than muxing, which we never actually used, it simply
transmits the file descriptors for the spawned process to whoever requests
them. if two links occur, it will be arbitrarily split between them, and
that's OK (because we only ever have one link).