set window size of tty to 80x24
I wrote a test for this but it was inexplicably flaky. gexec.Say would simply
not match, even though I could *see* the right output emitted through the
GinkgoWriter every time. The same test, without -tty, was fine.

If the test worked, it would simply be to execute 'tput cols; tput lines' and
check that it's '80\r\n24\r\n'.