fix accidental butchering of stdout/stderr
previously the same FD (master pty) was sent over for both stdout/stderr,
which meant linking would read from the same file in two separate goroutines,
printing to stdout/stderr, leaving the output jumbled and arbitrarily diced
and reordered.

instead, if we're spawning with a tty, send a dummy pipe for stderr, so that
only stdout will receive the data.