kill link processes hard
the default behavior for signal handling is to interrupt calls like
read(2) if they're reading from "slow devices". this meant that link
would see its read from stdin end with EOF, and try sending the eof
packet over to the spawn socket, causing the process to receive EOF,
which is not intended.

sigkill should be fine; the link process can die with fire.

Signed-off-by: Max Brunsfeld <mbrunsfeld@pivotallabs.com>