stream in from outside container's cgroup
otherwise we're prone to hit the container's memory limits, which is not
ideal as tar tends to use memory until it's flushed to disk.

stream in is a toplevel api call, so it's strange to have it hit this
issue, rather than be subject to the host's memory management.

now depends on nsenter, which it uses to enter the mount namespace and
stream in to the container's rootfs, as its vcap user.

[#77414660]

Signed-off-by: Max Brunsfeld <mbrunsfeld@pivotallabs.com>