initial wsh(d) work on shared volumes
API:

wshd now gets a --volumes argument, which points to a directory
that acts as a bridge to the host. volumes attached to the container
will appear here.

wsh now takes --bind-mount-name, --bind-mount-source, and
--bind-mount-destination flags. these configure a source on the host,
a destination in the container, and a common (unique) name for the
volume.

volumes can be attached on the fly, after a container is created. they
cannot currently be detached, and are always mounted read-write; these
are easy to add in the future.

implementation details:

a shared mount (mount --make-shared) is used to effectively mirror mount
trees across mount namespaces. each attached volume has a shared mount
point set up on the host, which appears under the container under a
canonical PATH/(name), which is then bind-mounted to the ultimate
destination in the container.

verbose example of the functionality:

  https://gist.github.com/vito/c2e20d0f579bedc30463

Signed-off-by: Johannes Petzold <jpetzold@pivotal.io>