Do mounts and networking config in host.
Once the root user is not always container root, it may not have
permissions to execute mounts and networking operations inside the
container. Therefore we have to do these operations in the host.

Also explicitly setuid and setgid to 0 so that we become container root
before running hook scripts.

We no longer have permissions to delete tmp/garden-host by the time the
container is created, so for now we're going to have to leave it.
See #83193088 which tracks figuring out how to delete this.

[#82838560]