Name bridges appropriately
The bridge names are prefixed with the Garden network interface
prefix, which consists of "w" followed by the tag.

This ensures that outbound packets from bridges are passed to the
filter forward chain by the following rule in bin/net.sh:

iptables -w -A FORWARD -i ${GARDEN_NETWORK_INTERFACE_PREFIX}+ \
    --jump ${filter_forward_chain}

Ensure bridge interface names are sufficiently short

There is a restriction that interface names are at most 15 bytes
long (at least for the stemcell we are using).

Also delete the bridge interface in destroy.sh. This is temporary.

[#80656918]

Signed-off-by: Steve Powell <spowell@pivotal.io>