WIP - chown UIDs/GIDs in to container namespace
We need to make sure that before we swap in to the user namespace the
UIDs in the rootfs file ownerships match the host-side UIDs we're going
to map the container users to. Otherwise as soon as we enter the
namespace users can't access their own files.

To make this work we parse through /etc/passwd inside the rootfs and for
every user we find we chown all files in the rootfs owned by that user
to the userid (and groupid) we're going to map that userid (groupid) to
inside the container.

- Also rename UIDPool to UIDBlockPool to reflect it now vending blocks
of uids

[#82838802]