Refactor iodaemon spawn
Break out utility functions.

Don't send PID back to caller as this is not needed and is confusing.

Don't close "unused" ends of pipes since this is distracting and the pipes are still in use, so nothing is really freed off.

Remove standard input parameter of spawn as this wasn't used.

Rename functions to reflect actual behaviour.

[#8326712]

Signed-off-by: Glyn Normington <gnormington@pivotal.io>