Ensure /etc/login.defs is present if needed.
Some useradd implementations fail if /etc/login.defs does not exist, so on failure, add the file and try again. This avoids adding the file when it is not needed which is inline with
Garden's philosophy of minimising changes to the rootfs.

Also, ensure /etc exists because Garden should cope with an empty rootfs.

Signed-off-by: Glyn Normington <gnormington@pivotal.io>