Close pipe ends which have been sent from the container to the host.
This avoids hangs where the host closes a pipe end and yet the pipe
remains open because of an unclosed pipe end in the container.

See, for example http://unixhelp.ed.ac.uk/CGI/man-cgi?pipe+7, which
says:

If all file descriptors referring to the write end of a pipe have  been
closed,	then  an attempt to read(2) from the pipe will see end-of-file
(read(2) will return 0).

[#90094414]