Prevent kill -9 causing a hang.
The streaming wait group causes a deadlock if the job is killed -9, in
spite of defers which close the job's ends of the stdout and stderr
spawner pipes.

The only solution seems to be to avoid waiting for streaming to
complete if the job terminates with exit status 255, which usually
indicates an error or kill -9.