Avoid deadlock when a process is killed.
Add an integration test of wsh and daemon.

Fix the round-trip test to use pipes rather than files.

Signed-off-by: George Lestaris <glestaris@pivotal.io>

Allow more streaming after a job has exited abnormally.

This is the best we can do with the current wsh/wshd interface. A
better solution involves making signalling a daemon function (part of
doing away with PID files) so that the daemon can know for sure
whether a process has been killed rather than relying on its exit
status being 255. Then the daemon could allow a bit more streaming to
occur if and only if the process has been killed (which is just being
helpful given the race between streamed output and the kill signal).

The only real downside of the current approach is that it will not
wait indefinitely for streaming to complete if a process is not killed
but exits with exit status 255. This is arguably an edge case anyway,
so the solution should suffice for now.

[#90094414]