Use timeout instead of polling to collect process output
In most cases this waits only as long as necessary for the process
streams to close. In the case where the process has spawned a child
which is still writing to stdout/err, waits up to 100 millieconds before
giving up.

This implementation has the advantage of being significantly more
idiomatic and simple (and the disadvantage of involving a timeout, and
therefore sometimes waiting longer than necessary in the fairly unusual case
where the process spawned a child which is still running when it exits).