Add wrapper around docker graph and graphdriver
Previously various parts of the code-base accessed docker.Graph and
docker.GraphDriver directly, which led to problems where
docker.GraphDriver was directly called and docker.Graph was out of sync
(for example). This refactor consolidates a single interface with a
smaller surface area to be used by all parts of the code that need a
graph-like layered filesystem manager.

This change also moves the graph cleanup logic in to the new 'layercake' package.

[#96458012]

Signed-off-by: Gareth Smith <gsmith@pivotal.io>