Ping returns an error if the graph directory is read-only.
This enables the health of BTRFS to be checked. (If BTRFS explodes, it
sets the graph directory read-only.)

The health check used by ping is located in the sysinfo package,
although it may be appropriate to separate this out later if it grows
or needs to be augmented with other health checks. This seemed better
than adding yet another argument to linux_backend.New.

The health check of BTRFS simply writes a file into the graph
directory. If the graph directory is invalid or refers to a file, this
also fails with suitable diagnostics, so there is no need for extra
logic or tests for these cases.

The hand-coded sysinfo.Provider mock is replaced with one generated by
counterfeiter, for ease of maintenance.

[#103614946]