Ping returns an error if the graph directory can't be written
This enables the health of BTRFS to be checked. (If BTRFS explodes, it
sets the graph directory read-only.)

The health check of BTRFS simply writes a file into the graph directory.
If this fails for any reason, an unrecoverable error is returned (in
practice, this should only occur if btrfs has hit problems or we've run
out of disk, neither of which is a situation that can be recovered from
at the moment).

[#103614946]

Signed-off-by: Julian Friedman <julz.friedman@uk.ibm.com>
Signed-off-by: Gareth Smith <gsmith@pivotal.io>