Make all graph operations go through the Coordinator
Allows the coordinator to easily lock writes and deletes against the
graph. This means we can now remove the error-prone Retainer logic.

The new structure has a single rootfsprovider.Cakeordinator at the top.
This uses a repository fetcher to fetch layers, and a layer creator to
add container layers (and namespace layers) on top. All access to the
graph now goes through the coordinator, which manages the interactions
and locking.

Both the fetchers and layer creators can create layers in the cake,
which is always safe, but only the coordinator can delete, and it
ensures no creation operations are running before doing so.

Signed-off-by: Gareth Smith <gsmith@pivotal.io>