Fix tests for latest Concourse
The integration test runner was causing the VM to hang when running in
latest Concourse on Linux kernel version 3.19.0-33-generic. We are
now not deleting the empty `/tmp/aufs_mount/node-X` directories of each test
(used for the graph directory of the Ginkgo node) in order to get around the
problem. Next test reuses these directories so we are not leaking them.
More info can be found in the story.

A similar workaround is enforced in the `container/system` unit tests.

[#108870596]

Signed-off-by: Georgi Sabev <georgethebeatle@gmail.com>