Remove destination filter from MASQUERADE rules
When using DNAT rules to forward traffic for the
external IP of the garden host into one of the
containers, this now allows internal containers to
talk to the external address of that service successfully,
when that service is on the same subnet of the
internal container originating the request.

This should retain the features of MASQUERADE required
for multi-homed garden hosts, and not change the behavior
of direct intra-subnet communication for garden containers, or
direct communication across subnets for garden containers.