don't run iodaemon in the background using bash
this means more threads but will allow us to get the exit status of the
iodaemon. otherwise, iodaemon will become an orphan process and the exit
status will be lost in init. Note: we can start one goroutine that calls
syscall.Wait4 but i don't see a reason to do so.

Signed-off-by: Brandon Shroyer <bshroyer@pivotal.io>